Move createLocalSocket() from Valetudo to MiioVacuum.
