+++ lib/Valetudo.js
@@ -1,7 +1,4 @@
-
-
 const fs = require("fs");
-const dgram = require("dgram");
 const DEFAULT_MAP = require("./res/default_map");
 const Webserver = require("./webserver/WebServer");
 const MqttClient = require("./MqttClient");
@@ -11,7 +8,6 @@ const {EventEmitter} = require("events");
 const SSHManager = require("./SSHManager");
 const Model = require("./miio/Model");
 const Logger = require("./Logger");
-const MiioSocket = require("./miio/MiioSocket");
 const Viomi = require("./devices/Viomi");
 const Roborock = require("./devices/Roborock");
 
@@ -38,7 +34,6 @@ const Valetudo = function() {
     this.vacuum;
 
     const robotArgs = {
-        localSocket: createLocalSocket(this.tokenProvider(), this.address),
         events: this.events,
         ip: this.address,
         configuration: this.configuration,
@@ -76,22 +71,6 @@ const Valetudo = function() {
         });
     }
 };
-
-function createLocalSocket(token, address) {
-    const socket = dgram.createSocket("udp4");
-    socket.bind();
-
-    return new MiioSocket({
-        socket: socket,
-        token: token,
-        onMessage: null,
-        deviceId: undefined,
-        rinfo: {address: address, port: MiioSocket.PORT},
-        timeout: undefined,
-        onConnected: undefined,
-        name: "local"
-    });
-}
 
 /**
  * Returns a {did: …, key:…, vendor: …, mac: …, model: …} object.
