+++ lib/devices/MiioVacuum.js
@@ -1,5 +1,7 @@
+const dgram = require("dgram");
 const Dummycloud = require("../miio/Dummycloud");
 const Logger = require("../Logger");
+const MiioSocket = require("../miio/MiioSocket");
 const RetryWrapper = require("../miio/RetryWrapper");
 const Status = require("../miio/Status");
 
@@ -14,7 +16,6 @@ const Status = require("../miio/Status");
  * @param options.deviceId {string} The unique Device-id of your robot
  * @param options.events {import("events").EventEmitter}
  * @param options.ip {string} Remote IP
- * @param options.localSocket {import("../miio/MiioSocket")}
  * @param options.tokenProvider {()=>Buffer}
  *
  * @constructor
@@ -33,7 +34,9 @@ const MiioVacuum = function(options) {
         onConnected: () => this.onCloudConnected(),
         onMessage: () => this.onMessage
     });
-    this.localRetryWrapper = new RetryWrapper(options.localSocket, options.tokenProvider);
+
+    this.localRetryWrapper = new RetryWrapper(
+        createLocalSocket(options.tokenProvider(), options.ip), options.tokenProvider);
 
     // Poll status once per minute.
     let pollStatusLoop = function() {
@@ -43,6 +46,22 @@ const MiioVacuum = function(options) {
     }.bind(this);
     this.pollStatus(pollStatusLoop);
 };
+
+function createLocalSocket(token, address) {
+    const socket = dgram.createSocket("udp4");
+    socket.bind();
+
+    return new MiioSocket({
+        socket: socket,
+        token: token,
+        onMessage: null,
+        deviceId: undefined,
+        rinfo: {address: address, port: MiioSocket.PORT},
+        timeout: undefined,
+        onConnected: undefined,
+        name: "local"
+    });
+}
 
 MiioVacuum.FAST_MAP_PULLING_STATES = [
     "REMOTE_CONTROL_ACTIVE", "CLEANING", "RETURNING_HOME", "MANUAL_MODE", "SPOT_CLEANING",
